requirements.txt
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.0
docker>=6.0.0
pandas>=2.0.0
networkx>=3.0.0
tqdm>=4.65.0
python-dotenv>=1.0.0
tabulate>=0.9.0
PyYAML>=6.0.0 